chore: Remove old snoozeDateParser in favor of multi-lang natural language parser
