+++ app/javascript/dashboard/helper/snoozeDateParser/index.js
@@ -1,0 +1,12 @@
+/**
+ * snoozeDateParser — Natural language date/time parser for snooze.
+ *
+ * Barrel re-export from submodules:
+ *   - parser.js: core parsing engine (parseDateFromText)
+ *   - localization.js: multilingual suggestion generator (generateDateSuggestions)
+ *   - suggestions.js: compositional suggestion engine
+ *   - tokenMaps.js: shared token maps and utility functions
+ */
+
+export { parseDateFromText } from './parser';
+export { generateDateSuggestions } from './localization';
